pescaria-chomik/application/requirements.txt
discord.py	
python-dotenv
